make a common setup-function
