+++ .github/workflows/dispatch-k6-performance.yml
@@ -48,7 +48,6 @@ on:
           - 'tests/k6/tests/enduser/performance/enduser-search.js'
           - 'tests/k6/tests/graphql/performance/graphql-search.js'
           - 'tests/k6/tests/serviceowner/performance/create-transmissions.js'
-          - 'tests/k6/tests/enduser/enduserSearchWithThresholds.js'
 
 run-name: ${{ inputs.tag }} vus ${{ inputs.vus }} duration ${{ inputs.duration }}
 jobs:
