+++ tests/k6/common/token.js
@@ -39,9 +39,9 @@ export function fetchToken(url, tokenOptions, type) {
     if (__VU == 0) {
       console.info(`Fetching ${type} token from token generator during setup stage`);
     }
-    // else {
-    //   console.info(`Fetching ${type} token from token generator during VU stage for VU #${__VU}`);
-    // }
+    else {
+      console.info(`Fetching ${type} token from token generator during VU stage for VU #${__VU}`);
+    }
     
     let response = http.get(url, tokenRequestOptions);
