+++ tests/k6/tests/serviceowner/performance/create-remove-dialog.js
@@ -6,6 +6,7 @@ import exec from 'k6/execution';
 import { getDefaultThresholds } from '../../performancetest_common/getDefaultThresholds.js';
 import { serviceOwners, endUsersPart } from "../../performancetest_common/readTestdata.js";
 import { createAndRemoveDialog } from '../../performancetest_common/createDialog.js';
+export { setup as setup } from '../../performancetest_common/readTestdata.js';
 
 const traceCalls = (__ENV.traceCalls ?? 'false') === 'true';
 
@@ -16,15 +17,6 @@ export let options = {
       'remove dialog'
     ])
 };
-
-export function setup() {
-  const totalVus = exec.test.options.scenarios.default.vus;
-  let parts = [];
-  for (let i = 1; i <= totalVus; i++) {
-      parts.push(endUsersPart(totalVus, i));
-  }
-  return parts;
-}
 
 export default function(data) {
   const myEndUsers = data[exec.vu.idInTest - 1];
