+++ tests/k6/tests/serviceowner/performance/serviceowner-search.js
@@ -1,10 +1,10 @@
-import { randomItem } from 'https://jslib.k6.io/k6-utils/1.4.0/index.js';
+import exec from 'k6/execution';
 import { serviceownerSearch } from '../../performancetest_common/simpleSearch.js'
 import { getDefaultThresholds } from '../../performancetest_common/getDefaultThresholds.js';
-import { serviceOwners, endUsers } from '../../performancetest_common/readTestdata.js';
+import { serviceOwners } from '../../performancetest_common/readTestdata.js';
+export { setup as setup } from '../../performancetest_common/readTestdata.js';
 
 const tag_name = 'serviceowner search';
-const isSingleUserMode = (__ENV.isSingleUserMode ?? 'false') === 'true';
 const traceCalls = (__ENV.traceCalls ?? 'false') === 'true';
 
 export let options = {
@@ -18,19 +18,10 @@ export let options = {
         'get transmissions',
         'get transmission'])
 };
-
-export function setup() {
-    const totalVus = exec.test.options.scenarios.default.vus;
-    let parts = [];
-    for (let i = 1; i <= totalVus; i++) {
-        parts.push(endUsersPart(totalVus, i));
-    }
-    return parts;
-  }
   
-  export default function(data) {
+export default function(data) {
     const myEndUsers = data[exec.vu.idInTest - 1];
     const ix = exec.vu.iterationInInstance % myEndUsers.length;
-    serviceownerSearch(serviceOwner, myEndUsers[ix], tag_name, traceCalls);
-  }
+    serviceownerSearch(serviceOwners[0], myEndUsers[ix], tag_name, traceCalls);
+}
